pyproject.toml
[tool.poetry]
authors = ["Ziyi Xi <ziyixi@mail.ustc.edu.cn>", "Weiqiang Zhu <wayne.weiqiang@gmail.com>"]
description = "PhaseNet-TF: Advanced Seismic Arrival Time Detection via Deep Neural Networks in the Spectrogram Domain, Leveraging Cutting-Edge Image Segmentation Approaches"
license = "MIT"
name = "phasenet-tf"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
hydra-core = "^1.3.2"
lightning = "~2.0"
numpy = "^1.24.2"
pandas = "^2.0.0"
pyrootutils = "^1.0.4"
python = ">=3.9,<3.12"
rich = "^13.3.4"
scipy = "^1.10.1"
segmentation-models-pytorch = "^0.3.2"
torch = "2.0.0"
nnAudio = "^0.3.2"

[tool.poetry.group.train]
optional = true

[tool.poetry.group.train.dependencies]
h5py = "^3.8.0"
hydra-colorlog = "^1.2.0"
hydra-submitit-launcher = "^1.2.0"
matplotlib = "^3.7.1"
obspy = "^1.4.0"
sqlalchemy = "<2.0"
wandb = "^0.14.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
httpx = "^0.24.0"
ipykernel = "^6.22.0"
joblib = "^1.2.0"
pylint = "^2.17.2"

[tool.poetry.group.api]
optional = true

[tool.poetry.group.api.dependencies]
fastapi = "0.86" # fix https://github.com/Lightning-AI/lightning/issues/17106 
pydantic = "^1.10.7" 
uvicorn = "^0.21.1" 
onnx = "^1.14.0"
onnxruntime = "^1.14.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
